+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Feature `TestEngine.GetTestFullName` to compare with `TestEngine.GetTestFullNames`.
 ### Fixed
 - Fix Multiple AssertionResults: `TestExecutionContextUtils.Clear();`
+### Updated
+- Update core `TestAssemblyService` and `TestService` to add fail tests using `TestInstance` method
 
 ## [1.0.9] / 2022-12-06
 ### Features